If redirecting, don't visit the home screen first
